chore: moved implementation of event bus to infra layer on shared context
